requirements.base.txt
--index-url https://download.pytorch.org/whl/cu118
--extra-index-url https://pypi.org/simple

# Critical ML stack (pinned for reproducibility)
torch==2.6.0
torchvision==0.21.0
transformers==4.46.3
tokenizers==0.20.3

# DeepSeek-OCR-2 runtime dependencies
einops>=0.7,<1
addict>=2.4,<3
easydict>=1.13,<2

# App/runtime
pillow>=10,<12
pypdfium2>=4,<5

# flash-attn build dependency (required for some build backends)
psutil>=5.9,<6
